webapp/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-docx==1.1.2
python-multipart==0.0.9
mammoth==1.8.0
pypdf==5.1.0
pdf2docx==0.5.8
docx2pdf==0.1.8
reportlab==4.2.5
weasyprint==62.3
